Transform list cleanup
