+++ enterprise/frontend/src/metabase-enterprise/transforms/components/TransformTable/TransformTable.module.css
@@ -1,0 +1,3 @@
+.row {
+  cursor: pointer;
+}
