+++ enterprise/frontend/src/metabase-enterprise/transforms/components/TransformTable/index.ts
@@ -1,0 +1,1 @@
+export * from "./TransformTable";
